requirements.txt
Django==3.1.14
django-jazzmin==2.2.9
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
django-cors-headers==3.5.0
drf-api-tracking==1.8.4
drf_generator==0.0.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dotenv==0.21.0
pytz==2024.2
sqlparse==0.5.1
django-filter==21.1
